wip creating custom styled components
