+++ dashboard/app/api/cloud/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { getAccessToken, withApiAuthRequired } from "@auth0/nextjs-auth0";
+import { getAccessToken } from "@auth0/nextjs-auth0";
 
 export async function POST(req) {
   try {
