+++ dashboard/components/tremor/index.tsx
@@ -1,0 +1,4 @@
+export * from "./card";
+export * from "./button";
+export * from "./button";
+export * from "./card";
