+++ dashboard/next.config.js
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  env: {
+    INTERCOM_APP_ID: process.env.INTERCOM_APP_ID,
+  },
+};
 
 module.exports = nextConfig;
