+++ dashboard/package.json
@@ -24,6 +24,7 @@
     "postcss": "8.4.30",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-use-intercom": "^5.1.4",
     "server-only": "^0.0.1",
     "swr": "^2.2.4",
     "tailwindcss": "3.3.3",
